refactored video page
