--- src/pages/video/Video.jsx
@@ -1,10 +1,0 @@
-const Video = () => {
-  return (
-    <div className='bg-neutral-500 h-128 aspect-video rounded-xl'>
-    </div>
-  );
-};
-
-
-
-export default Video;
